+++ lib/kitchen/selectors/standard_1.rb
@@ -10,8 +10,8 @@ module Kitchen
       #
       def initialize
         super
-        self.title_in_page              = "./*[@data-type = 'document-title']"
-        self.title_in_introduction_page = "./*[@data-type = 'document-title']"
+        self.title_in_page              = ".//*[@data-type='document-title']"
+        self.title_in_introduction_page = ".//*[@data-type='document-title']"
         self.page_summary               = 'section.summary'
         self.reference                  = '.reference'
         self.chapter                    = "div[data-type='chapter']"