+++ spec/kitchen_spec/page_element_spec.rb
@@ -4,12 +4,10 @@ require 'spec_helper'
 
 RSpec.describe Kitchen::PageElement do
 
-  let(:page_title_text) { 'A title!' }
-
   let(:metadata) do
     <<~HTML
       <div data-type="metadata" style="display: none;">
-        <h1 data-type="document-title" itemprop="name">#{page_title_text}</h1>
+        <h1 data-type="document-title" itemprop="name">Metadata title</h1>
       </div>
     HTML
   end
@@ -19,7 +17,7 @@ RSpec.describe Kitchen::PageElement do
       one_chapter_with_one_page_containing(
         <<~HTML
           #{metadata}
-          <div data-type="document-title">#{page_title_text}</div>
+          <div data-type="document-title">Page title</div>
         HTML
       )).pages.first
   end
@@ -35,27 +33,6 @@ RSpec.describe Kitchen::PageElement do
         </div>
       HTML
     ).pages.first!
-  end
-
-  let(:page_with_exercises) do
-    book_containing(html:
-      one_chapter_with_one_page_containing(
-        <<~HTML
-          <section data-depth="1" id="auto_m68764_fs-idm81325184" class="exercises">
-            <h3 data-type="title">Section Title</h3>
-            <div data-type="exercise" id="auto_m68764_fs-idm178529488">
-              <div data-type="problem" id="auto_m68764_fs-idp20517968">
-                <p id="auto_m68764_fs-idm197064800">Problem 1</p>
-              </div>
-            </div>
-            <div data-type="exercise" id="auto_m68764_fs-idm82765632">
-              <div data-type="problem" id="auto_m68764_fs-idp7685184">
-                <p id="auto_m68764_fs-idm164104512">Problem 2</p>
-              </div>
-            </div>
-          </section>
-        HTML
-      )).pages.first
   end
 
   let(:introduction_page) do
@@ -88,13 +65,18 @@ RSpec.describe Kitchen::PageElement do
       let(:metadata) { '' }
 
       it 'finds the title' do
-        expect(page1.title.text).to eq page_title_text
+        expect(page1.title.text).to eq 'Page title'
       end
     end
 
     context 'with metadata' do
-      it 'finds the title' do
-        expect(page1.title.text).to eq page_title_text
+      it 'skips metadata to find title' do
+        expect(page1.title.text).to eq 'Page title'
+      end
+
+      it 'errors when no title' do
+        page1.title.trash
+        expect { page1.title }.to raise_error(/Title not found for page/)
       end
     end
   end
@@ -130,11 +112,4 @@ RSpec.describe Kitchen::PageElement do
   it 'returns the metadata' do
     expect(page1.metadata).to match_normalized_html(metadata)
   end
-
-  describe '#exercises' do
-    it 'returns the exercises element' do
-      expect(page_with_exercises.exercises).to match_snapshot_auto
-    end
-  end
-
 end