+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Bake `chapter outline` in `neuroscience`
+* Add Shellcheck to Dockerfile
+* Fixup `PageElement`, add specs for `CompositePageElement`
+* Bake `answer key` in `lifespan-development`
+* Bake `notes` in `lifespan-development`
+* Bake `index` in `lifespan-development`
+* Bake `eoc` sections in `lifespan-development`
+* Bake `tables` in `lifespan-development`
 * Bake `figures` in `lifespan-development`
 * Create basic files for `lifespan-development` recipe